requirements.txt
pyqt5
vk_api
Pillow
playsound==1.2.2